Quiz API request and loading state
