+++ frontend/package.json
@@ -19,9 +19,11 @@
     "react-hook-form": "^7.18.0",
     "react-i18next": "^11.13.0",
     "react-icons": "^4.3.1",
+    "react-loading-skeleton": "^3.0.1",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "react-super-responsive-table": "^5.2.0",
+    "swr": "^1.0.1",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
